fix breaking test "use of const in strict mode" in node 0.12.x
There was a minor update to tap that introduces an update to a child dependency
which breaks compatibility with node 0.12.x when running `npm test`. By changing
the `^` to a `~` we lock down the dependency tree to only allow for patch
updates.

This fixes the failing travis build for pr npm/copy-concurrently/pull/2